requirements.txt
Django==1.11.4
djangorestframework==3.6.4
mysqlclient==1.3.12
olefile==0.44
Pillow==4.2.1
pytz==2017.2
